requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
freetype-py==2.1.0.post1
gif2numpy==1.3
idna==2.9
kaitaistruct==0.8
mss==5.0.0
numpy==1.18.2
numpy2gif==1.0
opencv-python==4.2.0.34
peewee==3.13.2
Pillow==7.0.0
psycopg2-binary==2.8.5
requests==2.7.0
soupsieve==2.0
urllib3==1.25.8